handla and add access token

+++ src/api/products/deleteProduct.js
@@ -1,14 +1,38 @@
 import axios from "axios";
+import { refreshAccessToken } from "../auth/refreshToken";
 
-export const deleteProduct = async ({ id, token }) => {
-  const {data} = await axios.delete(
-    `https://api.dahbelarby.com/api/products/${id}`,
-    {
-      headers: {
-        Authorization: `Bearer ${token}`,
-      },
+export const deleteProduct = async ({ id }) => {
+  let accessToken = localStorage.getItem("accessToken");
+
+  try {
+    const { data } = await axios.delete(
+      `https://api.dahbelarby.com/api/products/${id}`,
+      {
+        headers: {
+          Authorization: `Bearer ${accessToken}`,
+        },
+      }
+    );
+
+    return data.data;
+  } catch (error) {
+    if (error.response?.status === 401) {
+      const tokens = await refreshAccessToken();
+
+      accessToken = tokens.accessToken;
+
+      const { data } = await axios.delete(
+        `https://api.dahbelarby.com/api/products/${id}`,
+        {
+          headers: {
+            Authorization: `Bearer ${accessToken}`,
+          },
+        }
+      );
+
+      return data.data;
     }
-  );
 
-  return data.data;
+    throw error;
+  }
 };
